safeshare/requirements.txt
annotated-types==0.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
boto==2.49.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
Django==4.2.5
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
dr-scaffold==2.1.2
drf-nested-routers==0.93.4
drf-yasg==1.21.7
grpcio==1.59.0
idna==3.4
inflect==7.0.0
inflection==0.5.1
install==1.3.5
isort==5.12.0
mysqlclient==2.2.0
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==1.7.1
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.5
